finish refactor for form block. relates to #504.
